requirements.txt
Django==5.1.6
djangorestframework==3.16.0
asgiref==3.8.1
numpy==2.2.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
sqlparse==0.5.3
zxing-cpp==2.3.0
pyusb==1.3.1
